Added title in the framedwindow's top bar
This is really basic, and is still quite buggy. There are many improvements to be made
such as not loading the font every time, and storing the window's name atom globally, but
it works.
In order to implement most improvements, I think it's better to reefractor the way the program
runs.