Reacting to frame buttons events to move the window
This is a way nicer way of dealing with mouse press and movement events, and seems
to work smoother. It also doesn't steal the events from the target window, meaning
the mouse should now work more or less correctly. Unfortunately, it doesn't work inside
a window, so we can no longer switch indow by clicking on them. This needs to be fixed.