Moved to XCB.
Moved away from Xlib and changed everything to use XCB.

I was running into too much trouble trying to get Xlib to do what I
wanted, so I just… threw it away!

Well, it's broken as of now. It builds, but segfaults on startup, which
is not great, but most of the startup still seems to work, it seems to
break in the event loop, so I suspect that when events will be handeled
properly, everything will be fine.

Right now there still is lots of commented Xlib code, as not everything
is implemented with XCB yet, and I will be using this as a reference to
write the XCB code.